Added /login and /register routes.
