+++ client/package.json
@@ -15,6 +15,7 @@
     "@vitejs/plugin-react": "^5.2.0",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
+    "react-router-dom": "^7.8.2",
     "tailwindcss": "^4.2.1",
     "vite": "^7.3.1"
   },
